Controls sort of working.
Had a lot of trouble in the beginning that I suspect were due to
mount/unmount problems.  Refactored to make the polytope class
immutable.  Any caching is done on the transformStack.  Simplices are
having serious problems.

I think several other problems are due to events and rendering being
asynchronous.  Or something.