fix the infamous simplex bug.
It turned out that when we called updateDimensions from the UI, we were
passing this.event.value to setDimensions, and frion there to
getPolytopesFor, without converting it from a string to a number.
Thanks to JavaScript's "helpful" type coercions, this was causing the
number of vertices in a simplex, which is dim+1, to _also_ be a string.

It didn't make a difference anyplace but the simplex, because in the
cube and orthoplex the number of vertices is given by an expression that
involves multiplication, which forces the coercion at the call site.