Fix simplex.
Getting the vertices of the simplex right turns out to have been
difficult, and I'm still not sure it's right -- you have to test for
onUnitSphere with 0 digits of precision -- one of them ends up at a norm
of .93something.  There is probably still something wrong, but it looks
close enough for filk.