Fix simplex edge construction
It was constructing edges and vertices in a local array, then assigning
this.vertices=vertices and this.edges=edges.  Except the assignment of
this.edges was missing.

There's no need to use a local, so push the vertices and edges directly
into this.vertices and this.edges, which is what all the other polytopes
do.